Number of fixes and prep for new version 4.9.0
   - SQL Prep for new version
       - Merged all changes to date into one version script
       - Moved old version scripts into their own folders
       - Added fix for stats title count on the date range search (UF)
       - Still need to create new complete.sql for new install
   - Corrected upload of JPEGs behavior
        - Thumbnails generated from uploaded JPEGs now
        - Large JPEGs are saved, converted to TIFF, and then smaller versions made.
        - Builder will then generate the JPEG2000s
    - Corrected issue deleting existing items (MBDA-14)
    - Corrected issue where the host administrator check box was not saving to the database
    - Corrected the styling on the private items report when more than one page of private titles
         - Buttons use correct styling
         - Number of titles displayed ( i.e., 1 - 20 of 45 matching titles ) now shown
     - Updated QC exception catching to work correctly and to email the error in when it occurs (WOLF-29)
     - Builder image derivative creation fixes
         - Was not checking for JPEG2000 missing, only JPEGs.  Now correctly checks JPEG2000s as well. (USF)
         - No longer copies the TIFF to a local temporary folder for processing.  All processing happens in situ (SOBEK-51/81)
      - Corrected tab direction on the usage reports
      - Printing citation online now works correctly again (WOLF-34)
      - Corrected HTML facet link for donors in the citation - "( donor )" was messing it up (WOLF-30)
      - Corrected issue that DARK items were a little too resilient in loading downloads and images and were semi-accessible via the TOC (UOC-8)
      - Corrected issue preventing user group changes from correctly saving (SOBEK-66/SOAS)
      - When deleting a PDF in the online interface, the related text and thumbnail files are also deleted (SOBEK-52/USF)
      - Verified that downloads with accents works just fine throughout the system (SOBEK-47)

+++ SobekCM_Engine_Library/Items/SobekCM_METS_Based_ItemBuilder.cs
@@ -638,11 +638,14 @@ namespace SobekCM.Engine_Library.Items
 
 			pageseq = 0;
 			List<Page_TreeNode> pages_encountered = new List<Page_TreeNode>();
-			foreach (abstract_TreeNode rootNode in Package_To_Finalize.Divisions.Physical_Tree.Roots)
-			{
-				recurse_through_nodes(Package_To_Finalize, rootNode, pages_encountered);
-			}
-			Package_To_Finalize.Web.Static_PageCount = pages_encountered.Count;
+		    if (!Package_To_Finalize.Behaviors.Dark_Flag)
+		    {
+		        foreach (abstract_TreeNode rootNode in Package_To_Finalize.Divisions.Physical_Tree.Roots)
+		        {
+		            recurse_through_nodes(Package_To_Finalize, rootNode, pages_encountered);
+		        }
+		    }
+		    Package_To_Finalize.Web.Static_PageCount = pages_encountered.Count;
 			Package_To_Finalize.Web.Static_Division_Count = divseq;
 
 			// Make sure no icons were retained from the METS file itself
@@ -942,7 +945,7 @@ namespace SobekCM.Engine_Library.Items
 			}
 
 			// Step through each download and make sure it is fully built
-			if (Package_To_Finalize.Divisions.Download_Tree.Has_Files)
+			if (( !Package_To_Finalize.Behaviors.Dark_Flag ) && ( Package_To_Finalize.Divisions.Download_Tree.Has_Files))
 			{
 				string ead_file = String.Empty;
 				int pdf_download = 0;
@@ -1086,14 +1089,14 @@ namespace SobekCM.Engine_Library.Items
 			}
 
 			// Look for the HTML type views next, and possible set some defaults
-			if (viewsFromDb.ContainsKey(View_Enum.HTML))
+            if ((!Package_To_Finalize.Behaviors.Dark_Flag) && viewsFromDb.ContainsKey(View_Enum.HTML))
 			{
 				Package_To_Finalize.Behaviors.Add_View(viewsFromDb[View_Enum.HTML]);
 				viewsFromDb.Remove(View_Enum.HTML);
 			}
 
 			// Copy the TEI flag
-			if (viewsFromDb.ContainsKey(View_Enum.TEI))
+            if ((!Package_To_Finalize.Behaviors.Dark_Flag) && viewsFromDb.ContainsKey(View_Enum.TEI))
 			{
 				Package_To_Finalize.Behaviors.Add_View(viewsFromDb[View_Enum.TEI]);
 				viewsFromDb.Remove(View_Enum.TEI);
@@ -1117,7 +1120,7 @@ namespace SobekCM.Engine_Library.Items
 			}
 
 			// Finally, add all the ITEM VIEWS
-			if ((Package_To_Finalize.Web.Pages_By_Sequence != null) && (Package_To_Finalize.Web.Pages_By_Sequence.Count > 0))
+            if ((!Package_To_Finalize.Behaviors.Dark_Flag) && (Package_To_Finalize.Web.Pages_By_Sequence != null) && (Package_To_Finalize.Web.Pages_By_Sequence.Count > 0))
 			{
                 // Look for the RELATED IMAGES view next
                 if (viewsFromDb.ContainsKey(View_Enum.RELATED_IMAGES))
